Added automatic stowing and directory checks

+++ .scripts/install-nvim.sh
@@ -2,7 +2,7 @@
 
 # script for setting up neovim to use .vimrc in home directory, which symbolically links to the dotfiles git repo
 
-# ensure script is not run with sudo so current user can be identified
+# ensure script is not run with sudo so current user and home can be identified
 if [ "$EUID" == "0" ]; then
 	echo "Please run this script without sudo"
 	echo "$0"
@@ -11,30 +11,58 @@ fi
 
 home=$HOME
 
-# TODO: check if ~/.dotfiles exists
+# check that ~/.dotfiles exists for the current user
+if [ ! -d "$home/.dotfiles" ]; then
+	echo "Please install dotfiles repo at ~/.dotfiles"
+	exit 1
+fi
 
-# installing neovim
-echo "INFO: Installing neovim from source"
+# ensure that script is run from .scripts directory
+if [ "$PWD" != "$home/.dotfiles/.scripts" ]; then
+	echo "Please run this script from within ~/.dotfiles/.scripts"
+	exit 1
+fi
+
+# installing neovim dependencies
+echo "INFO: Installing neovim dependencies"
 sudo apt-get update
 sudo apt-get remove -y neovim
-	sudo apt-get install -y ninja-build gettext libtool libtool-bin autoconf automake cmake g++ pkg-config unzip
-	sudo apt-get install -y python python-dev python3 python3-dev python-pip
+sudo apt-get install -y ninja-build gettext libtool libtool-bin autoconf automake cmake g++ pkg-config unzip
+sudo apt-get install -y python-dev python-pip python3-dev python3-pip
+pip install --user pynvim
+sudo apt-get install -y stow
+echo "INFO: Installing neovim dependencies DONE"
 
-	mkdir -p /usr/local/stow
+# remove neovim directory if it exists in home directory
+if [ -d "$home/neovim" ]; then
+	rm -rf $home/neovim
+fi
 
-# TODO check if neovim folder exists
-# TODO check if stow folder exists
+# create stow directory if it does not exist
+if [ ! -d "/usr/local/stow" ]; then
+	sudo mkdir -p /usr/local/stow
+fi
 
-	git clone https://github.com/neovim/neovim $home/neovim
-	cd $home/neovim
-	git checkout stable
-	rm -rf build
-	make CMAKE_BUILD_TYPE=RelWithDebInfo
-	sudo make CMAKE_INSTALL_PREFIX=/usr/local/stow/nvim install
-	pip install --user neovim
-echo "INFO: Installing neovim DONE"
+# installing neovim from source
+echo "INFO: Installing neovim from source"
+git clone https://github.com/neovim/neovim $home/neovim
+cd $home/neovim
+rm -rf build
+make CMAKE_BUILD_TYPE=RelWithDebInfo CMAKE_INSTALL_PREFIX=/usr/local/stow/nvim
+sudo make install
+echo "INFO: Installing neovim from source DONE"
 
-# TODO symlink .vimrc and .config/nvim using stow
+# symlinking neovim binary files
+echo "INFO: Symlinking neovim bin files"
+cd /usr/local/stow
+sudo stow --verbose=2 nvim
+echo "INFO: Symlinking neovim bin files DONE"
+
+# symlinking .vimrc and init.vim files
+echo "INFO: Symlinking .vimrc and init.vim files"
+cd $home/.dotfiles
+stow --verbose=2 vim
+echo "INFO: Symlinking .vimrc and init.vim files DONE"
 
 # add neovim to alternatives list
 echo "INFO: Updating default editors"
@@ -43,17 +71,9 @@ echo "INFO: Updating default editors"
 	sudo update-alternatives --install /usr/bin/editor editor /usr/local/bin/nvim 60
 echo "INFO: Updating default editors DONE"
 
-# create init.vim file which points to .vimrc in home directory
-#echo "INFO: Configuring init.vim"
-#mkdir -p $home/.config/nvim
-#echo "set runtimepath^=~/.vim runtimepath+=~/.vim/after" > $home/.config/nvim/init.vim
-#echo "let &packpath = &runtimepath" >> $home/.config/nvim/init.vim
-#echo "source ~/.vimrc" >> $home/.config/nvim/init.vim
-#echo "INFO: Configuring init.vim DONE"
-
 # install vim-plug and all plugins in .vimrc
 echo "INFO: Installing vim plugins"
-vim +"PlugInstall | q! | q!" ~/$RANDOM.vim --headless
+vim +"PlugInstall | q! | q!" ~/$RANDOM.txt --headless
 echo "INFO: Installing vim plugins DONE"
 
 # setting up vim-airline
